Rename 'ml.infer_trained_model_deployment' to 'ml.infer_trained_model'
