+++ specification/_json_spec/ml.infer_trained_model.json
@@ -1,0 +1,56 @@
+{
+  "ml.infer_trained_model": {
+    "documentation": {
+      "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/infer-trained-model.html",
+      "description": "Evaluate a trained model."
+    },
+    "stability": "experimental",
+    "visibility": "public",
+    "headers": {
+      "accept": ["application/json"],
+      "content_type": ["application/json"]
+    },
+    "url": {
+      "paths": [
+        {
+          "path": "/_ml/trained_models/{model_id}/_infer",
+          "methods": ["POST"],
+          "parts": {
+            "model_id": {
+              "type": "string",
+              "description": "The unique identifier of the trained model.",
+              "required": true
+            }
+          }
+        },
+        {
+          "path": "/_ml/trained_models/{model_id}/deployment/_infer",
+          "methods": ["POST"],
+          "parts": {
+            "model_id": {
+              "type": "string",
+              "description": "The unique identifier of the trained model.",
+              "required": true
+            }
+          },
+          "deprecated": {
+            "version": "8.3.0",
+            "description": "/_ml/trained_models/{model_id}/deployment/_infer is deprecated. Use /_ml/trained_models/{model_id}/_infer instead"
+          }
+        }
+      ]
+    },
+    "params": {
+      "timeout": {
+        "type": "time",
+        "required": false,
+        "description": "Controls the amount of time to wait for inference results.",
+        "default": "10s"
+      }
+    },
+    "body": {
+      "description": "The docs to apply inference on and inference configuration overrides",
+      "required": true
+    }
+  }
+}
